Updated Brightspace handler and fixed assignment bugs

+++ IguideME.Web/Models/App/AppAssignment.cs
@@ -29,7 +29,7 @@ namespace IguideME.Web.Models.App
         public bool Muted { get; set; }
 
         [JsonProperty("due_date")]
-        public int DueDate { get; set; }
+        public float DueDate { get; set; }
 
         [JsonProperty("max_grade")]
         public double MaxGrade { get; set; }
@@ -43,7 +43,7 @@ namespace IguideME.Web.Models.App
             string title,
             bool published,
             bool muted,
-            int dueDate,
+            float dueDate,
             double maxGrade,
             AppGradingType gradingType)
         {
